Upadte P4Runtime related resources
In particular update links to p4lang/PI to p4lang/p4runtime (which is
where the spec is now hosted) when appropriate.
The "p4-runtime" web page is the first result when one googles for
"P4Runtime", but it was mostly obsolete so I did some edits. I'm fine
with removing the page altogether if it is the right thing to do.